requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
SQLAlchemy==2.0.23
typing_extensions==4.9.0
Werkzeug==3.0.1
